Enable tag builds to be automatically pushed to bintray
Inspiration taken from http://bithacker.org/post/120996678805/publishing-artifacts-to-bintray-from-travis-ci

Essentially, nothing has changed for normal branch or master builds. TravisCI builds of tags will be pushed to bintray with the tag as part of the version number.

There's probably a bunch of optimisation or alternatives, but this is pretty quick and simple.