Throwables.propagate is deprecated
Throwables.propagate is deprecated.  Google have written a nice wiki
page explaining in great detail why it's deprecated:

https://github.com/google/guava/wiki/Why-we-deprecated-Throwables.propagate

It's scheduled for removal.  We should stop using it.

In this case, we only ever caught it on exceptions which were known at
compile time to be checked exceptions, so we can just wrap with
RuntimeException directly.