Add AccessLogstashConsoleAppenderFactory
This adds an appender specifically for access logs.  It allows you to
have request logs emitted in JSON format.  To use it you add the
following to your dropwizard config:

    server:
      requestLog:
        - type: access-logstash-console

This will only work if you are using logback-access style request logs
-- this is the default, but it may be disabled if you have
`type: classic` in your requestLog configuration.

This won't change existing behaviour; it only adds the new appender.

I've only added console logging for the moment because I'm
specifically targetting apps in a PaaS which will log to console.  We
can add File or Syslog appenders later if there's a demonstrated need.