add Host header to output
spelled `host` in the JSON output for consistency with filebeat
nginx/apache2 plugins

The test only asserts `startsWith` rather than `isEqualTo` because the
full Host header in this test is localhost:31337 for some ephemeral
port number, which will change for each test run.  We could ask the
DropwizardAppRule for its port number and make a tighter assertion but
this feels like overkill.